ml_model/requirements.txt
pandas
flask
numpy
scikit-learn
azure-storage-blob